refactor: intValue 변환 제거

+++ src/test/java/roomescape/domain/reservation/ReservationApiTest.java
@@ -227,7 +227,7 @@ class ReservationApiTest {
                 .log()
                 .all()
                 .when()
-                .delete("/reservations/" + id.intValue())
+                .delete("/reservations/" + id)
                 .then()
                 .log()
                 .all()
